+++ rand_pcg/tests/lcg128cmdxsm64.rs
@@ -57,19 +57,14 @@ fn test_lcg128cmdxsm64_reference() {
 #[cfg(feature = "serde")]
 #[test]
 fn test_lcg128cmdxsm64_serde() {
-    use bincode;
-    use std::io::{BufReader, BufWriter};
+    use postcard;
 
     let mut rng = Lcg128CmDxsm64::seed_from_u64(0);
 
-    let buf: Vec<u8> = Vec::new();
-    let mut buf = BufWriter::new(buf);
-    bincode::serialize_into(&mut buf, &rng).expect("Could not serialize");
+    let buf = postcard::to_allocvec(&rng).expect("Could not serialize");
 
-    let buf = buf.into_inner().unwrap();
-    let mut read = BufReader::new(&buf[..]);
     let mut deserialized: Lcg128CmDxsm64 =
-        bincode::deserialize_from(&mut read).expect("Could not deserialize");
+        postcard::from_bytes(&buf).expect("Could not deserialize");
 
     for _ in 0..16 {
         assert_eq!(rng.next_u64(), deserialized.next_u64());